Refactor yamlpaster_test.go
